refactor: フォーマット関数・クレジット定数を共通ユーティリティに切り出し
formatTime, formatDuration, estimateCredits, CREDITS_PER_MINUTE を
utils/transcription.ts に集約し、3コンポーネントの重複コードを削除
